// elza/elza-react/src/api/generated/model/ae-item-bit-vo.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AeItemBitVOAllOf } from './ae-item-bit-voall-of';
import { AeItemVO } from './ae-item-vo';

/**
 * Hodnota atributu - boolean
 * @export
 * @interface AeItemBitVO
 */
export interface AeItemBitVO extends AeItemVO {
    /**
     * logický výraz
     * @type {boolean}
     * @memberof AeItemBitVO
     */
    value: boolean;
}